Rename and move recommenders_provider.py
