+++ BookClub/tests/test_recommenders.py
@@ -1,7 +1,6 @@
 from django.test import TestCase, tag
 from BookClub.models import User, BookReview
-from BookClub.recommender_module.recommenders.resources.popular_books_recommender_methods import PopularBooksMethods
-from BookClub.recommender_module.recommenders import get_recommendations
+from BookClub.recommender_module import recommendations_provider
 
 @tag('recommenders')
 class RecommendersTestCase(TestCase):
@@ -13,7 +12,7 @@ class RecommendersTestCase(TestCase):
 
 
     def test_get_popularity_recommendations(self):
-        recommendations = get_recommendations.get_popularity_recommendations(self.user)
+        recommendations = recommendations_provider.get_popularity_recommendations(self.user)
         self.assertEqual(len(recommendations), 10)
         self.assertTrue(all(isinstance(isbn, str) for isbn in recommendations))
 
@@ -24,5 +23,5 @@ class RecommendersTestCase(TestCase):
         pass
 
     def test_get_books_read_by_user_no_book(self):
-        user_read_books = get_recommendations.get_user_read_books(self.user)
+        user_read_books = recommendations_provider.get_user_read_books(self.user)
         self.assertEqual(user_read_books, [])
